requirements.txt
# Core Requirements
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic>=2.0
pydantic-settings>=2.0
email-validator>=2.0.0  # Required for email validation in Pydantic

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.28.0

# Authentication
passlib[bcrypt]==1.7.4  # For password hashing

# OpenAI & Embeddings
openai==1.12.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.3

# PyTorch (compatible versions)
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0

# Utilities
tqdm==4.66.1
loguru==0.7.2
requests==2.31.0
aiofiles==23.1.0  # For file handling
huggingface-hub==0.19.4  # For model downloads

# Testing
pytest==7.4.4
httpx==0.26.0
pytest-cov==4.1.0